// src/shared/interface/user-types.ts
export interface ChildSignupInterface {
    username: string;
    email: string;
    phoneNo: string;
    password: string;
}


export interface ParentSignupInterface {
    name: string;
    username: string;
    email: string;
    phoneNo: string;
    password: string;
    gender: string;
    occupation: string;
}